src/recommendations/requirements.txt
aniso8601==10.0.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-JWT-Extended==4.7.1
flask-marshmallow==1.3.0
Flask-SQLAlchemy==3.1.1
Flask-Cors==5.0.1
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-api-core==2.24.2
MarkupSafe==3.0.2
marshmallow==3.26.1
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==1.4.1
packaging==24.2
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.38
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
pytest==8.3.5
pytest-cov==6.0.0
pytest-mock==3.14.0
opencv-python-headless==4.11.0.86
temp==2020.7.2
openai==1.79.0
httpx==0.28.1